+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This changelog covers what's changed in the MPDX APIs.
 
+## 6 April 2017
+- Contacts Wildcard Search is now returning contacts where the value is in notes.
+- Addresses are now returning the source_donor_account relationship.
+- Contacts are now returning the primary_or_first_person relationship.
+
 ## 5 April 2017
 - Added total_donation to donation detail
 - Amount under donation endpoint is now just returning an amount without symbol